fix: 修复clippy警告，优化代码质量
修复内容:
- 移除未使用的导入 (warn, Duration, Instant, SocketAddr等)
- 使用strip_prefix替代手动字符串切片
- 简化if-else分支 (metrics.rs)
- 使用let语句替代match单分支绑定
- 使用结构体初始化语法替代字段赋值
- 定义类型别名简化复杂类型
- 添加测试中缺失的Instant导入

测试结果: 639个测试全部通过

+++ src/services/voice_service.rs
@@ -168,7 +168,7 @@ impl VoiceStorage {
         .bind(params.session_id.as_deref())
         .bind(&params.file_path)
         .bind(&params.content_type)
-        .bind(params.duration_ms as i64)
+        .bind(params.duration_ms)
         .bind(params.file_size)
         .bind(waveform_json)
         .bind(now)
@@ -176,7 +176,7 @@ impl VoiceStorage {
         .await?;
 
         let id: i64 = result.try_get("id")?;
-        self.update_user_stats(&params.user_id, params.duration_ms as i64, params.file_size)
+        self.update_user_stats(&params.user_id, params.duration_ms, params.file_size)
             .await?;
 
         Ok(id)
